Renewable Auto Dev/requirements.txt
# Core AI and Agent Framework
pydantic-ai==0.0.14
pydantic==2.8.2
openai==1.47.0

# Database and Vector Storage
supabase==2.7.4
vecs==0.4.7
pgvector==0.3.3
sqlalchemy==2.0.35
psycopg2-binary==2.9.9

# Document Processing and RAG
langchain==0.3.0
langchain-community==0.3.0
langchain-openai==0.2.0
sentence-transformers==3.1.1
pypdf2==3.0.1
python-docx==1.1.2
openpyxl==3.1.5
faiss-cpu==1.8.0

# External Integrations
dropbox==12.0.2
requests==2.32.3
httpx==0.27.2

# Environment and Configuration
python-dotenv==1.0.1
pydantic-settings==2.5.2
pyyaml==6.0.2

# Utilities and Processing
pandas==2.2.3
numpy==2.1.1
beautifulsoup4==4.12.3
markdownify==0.13.1

# Development and Testing
pytest==8.3.3
pytest-cov==5.0.0
pytest-asyncio==0.24.0
black==24.8.0
flake8==7.1.1
mypy==1.11.2

# Monitoring and Logging
loguru==0.7.2
rich==13.8.1

# Security
cryptography==43.0.1

# API and Web Framework (for future web interface)
fastapi==0.115.0
uvicorn==0.31.0

# Async Support
asyncio==3.4.3
aiofiles==24.1.0 